LayoutTests/fast/page-color-sampling/color-sampling-ignores-large-sticky-container-expected.txt

PASS colors.top is null
PASS colors.left is null
PASS colors.right is null
PASS colors.bottom is null
PASS successfullyParsed is true

TEST COMPLETE
Hello world
